Release/1.7.1 (#10)
* VST3 Signing; Fix Live for Mac (#2376)

Live for Mac checks a signature of its sub-entries and cmake
ejected an invalid one for surge. Correct this by signing
all our products with an empty signature after assembly.

Closes #2374

* Add "Zoom to Default" menu item (#2377)

Closes #2375

* Change back to C++-14 on mac/linux; and MacOS minimum 10.9 (#2380)

This change backs us down from C++-17 to C++-14 on mac/linux which allows us
to re-activate MacOS 10.9 rather than 10.12. It keeps us on C++-17 on windows
so VS2019 can work properly with filesystem.

Closes #2378

* A few small 1.7.1 issues (#2382)

- Improve the default skin error message. Closes #2379
- Remove some unused SSE1 code. Closes #2341

* Unison Absolute Sync on Classic and S&H oscillator fix (#2383)

Resetting the sync state on Classic and S&H ignored absolute
on unison voices. Fix by resetting t appropriately.

Closes #2365

* Fix a pointer-compare error (#2385)

We were checking the validity of a member variable, which
was true but generated a clang warning, which on FreeBSD
resulted in the compilation breaking. So activate that warning-as-error
on mac and fix it.

Closes #2384

* Activate Clang and Ubuntu 20 on Linux in the pipeline (#2388)

Run a pipeline build with Ubuntu20
Run a pipeline build with Clang-10 on Ubuntu20
Fix the problems with Clang-10 on Ubuntu20

Closes #2386

* More Clang Changes (#2390)

Clang complains about large ints auto-casting to float.
Moreover, clang changes the flag to supress that warning
from clang 9 -> clang 10.

So just put in explicit casts to float.

Addresses #2386

* Re-introduce the LV2 idlefactor (#2393)

Closes #2391

* Update README.md

* Remove the SFXB properly now our VST3s are in a subdir (#2398)

This, combined with a pipeline issue, caused the 1.7.0
64 bit windows SurgeEffectsBank VST3 to not work

* Attempt to fix StudioOne zoom failure problems (#2401)

StudioOne Zoom fails sometimes in the VST3 API and that sends
Surge into a bit of a death spiral. Try to exit that death spiral
by having that failure try to set zoom to 100%, but with a recursion
guard.

Addresses #2397

* Change the mac asset lookup semantics slightly (#2404)

Use the newer of ~/Library vs /Library. This means if you did a build
in 2018 and install 1.7.1 you get the right result.

Closes #2396

* Adapt number of filter subtype menu entries based on filter type (#2399)

* Add parameter formatting for filter subtype
Tweak parameter formatting for env shape

* Adapt number of filter subtype menu entries based on filter type

* Unique filter subtype stringification based on filter type
Don't show RMB context menu if subtype button is hidden

* Slight tweak of filter type/subtype/F2 offset mode names

* Move Linuxes back to C++-17 (#2406)

Address #2388

* If we are in C++17 mode on linux, check with C++17 for fs (#2407)

addresses #2386

Co-authored-by: Paul <baconpaul@users.noreply.github.com>
Co-authored-by: EvilDragon <mario.kruselj@gmail.com>